add test and RequestRunner

+++ ShippingBotServer/src/main/java/com/example/shippingbotserver/dao/DaoProcessing.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.Objects;
+import java.util.Optional;
 
 @Slf4j
 @Component
@@ -22,7 +23,8 @@ public class DaoProcessing {
     private final LoverShower shower;
 
     @Autowired
-    public DaoProcessing(LoverRepository repository, LoverModelRepository modelRepository, FormHandler formHandler, LoverShower shower) {
+    public DaoProcessing(LoverRepository repository, LoverModelRepository modelRepository,
+                         FormHandler formHandler, LoverShower shower) {
         this.repository = repository;
         this.modelRepository = modelRepository;
         this.formHandler = formHandler;
@@ -30,7 +32,10 @@ public class DaoProcessing {
     }
 
     public FormLoverModel findLoverById(Long request) {
-        Lover person = repository.findById(request).get();
+        Optional<Lover> op = repository.findById(request);
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover person = op.get();
         log.debug("log message{}", "person was received: " + person);
         drawPicture(person);
         log.debug("log message{}", "image was draw");
@@ -40,8 +45,11 @@ public class DaoProcessing {
         return form;
     }
 
-    public FormLoverModel getQuestionnaire(Long id){
-        Lover lover = repository.findById(id).get();
+    public FormLoverModel getQuestionnaire(Long id) {
+        Optional<Lover> op = repository.findById(id);
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover lover = op.get();
         log.debug("log message{}", "person was received: " + lover);
         LoverModel loverModel = shower.showMe(lover, modelRepository, repository);
         drawPicture(new Lover(loverModel));
@@ -53,9 +61,12 @@ public class DaoProcessing {
         return formLoverModel;
     }
 
-    public FormLoverModel getFavorite(Long id, String action){
+    public FormLoverModel getFavorite(Long id, String action) {
         boolean increment = action.equalsIgnoreCase("Вправо");
-        Lover lover = repository.findById(id).get();
+        Optional<Lover> op = repository.findById(id);
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover lover = op.get();
         log.debug("log message{}", "person was received: " + lover);
         LoverModel favorite = lover.getLover(increment);
         String status = shower.getLoverModelStatus(lover, favorite);
@@ -65,16 +76,16 @@ public class DaoProcessing {
         Lover res = new Lover(favorite);
         res.setName(formHandler.getName());
         FormLoverModel formLoverModel = new FormLoverModel(res);
-        try{
-            formLoverModel.initBytes();
-        }catch (Exception ignored){
-        }
+        formLoverModel.initBytes();
         formLoverModel.setStatus(status);
         return formLoverModel;
     }
 
-    public FormLoverModel dislike(Lover id){
-        Lover lover = repository.findById(id.getId()).get();
+    public FormLoverModel dislike(Lover id) {
+        Optional<Lover> op = repository.findById(id.getId());
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover lover = op.get();
         log.debug("log message{}", "person was received: " + lover);
         LoverModel loverModel = shower.showMe(lover, modelRepository, repository);
         if (Objects.equals(loverModel.getId(), -1L)) {
@@ -82,8 +93,11 @@ public class DaoProcessing {
         }
         lover.getDislikes().add(loverModel);
         repository.save(lover);
+
         log.debug("log message{}", "person was disliked" + loverModel);
+
         log.debug("log message{}", "person was updated");
+
         loverModel = shower.showMe(lover, modelRepository, repository);
         log.debug("log message{}", "Next person was received" + loverModel);
         Lover res = new Lover(loverModel);
@@ -93,7 +107,10 @@ public class DaoProcessing {
     }
 
     public FormLoverModel like(Lover id) {
-        Lover lover = repository.findById(id.getId()).get();
+        Optional<Lover> op = repository.findById(id.getId());
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover lover = op.get();
         log.debug("log message{}", "person was received: " + lover);
         LoverModel loverModel = shower.showMe(lover, modelRepository, repository);
         if (Objects.equals(loverModel.getId(), -1L)) {
@@ -101,14 +118,19 @@ public class DaoProcessing {
         }
         lover.getLike().add(loverModel);
         String status = "";
-        if (shower.getLoverModelStatus(lover, loverModel).equals("Взаимность")){
+        if (shower.getLoverModelStatus(lover, loverModel).equals("Взаимность")) {
             status = "Вы любимы: " + shower.translateLover(loverModel.getGender()) + ", " + loverModel.getName();
         }
         repository.save(lover);
+
         log.debug("log message{}", "person was liked" + loverModel);
+
         log.debug("log message{}", "person was updated");
+
         loverModel = shower.showMe(lover, modelRepository, repository);
+
         log.debug("log message{}", "Next person was received" + loverModel);
+
         Lover res = new Lover(loverModel);
         res.setName(formHandler.getName());
         FormLoverModel formLoverModel = getFormLoverModel(loverModel, status, res);
@@ -134,18 +156,31 @@ public class DaoProcessing {
     }
 
     public void deleteLover(Long id) {
-        Lover lover = repository.findById(id).get();
+        Optional<Lover> op = repository.findById(id);
+        if (!op.isPresent())
+            throw new RuntimeException("Пользователь не найден");
+        Lover lover = op.get();
+
         log.debug("log message{}", "person was received: " + lover);
+
         repository.delete(lover);
+
         log.debug("log message{}", "person was deleted");
     }
 
 
     public void update(Lover lover) {
-        Lover orig = repository.findById(lover.getId()).get();
-        log.debug("log message{}", "person was received: " + orig);
-        shower.loverUpdate(lover, orig);
+        Optional<Lover> op = repository.findById(lover.getId());
+        Lover orig;
+
+
+        if (op.isPresent()) {
+            orig = op.get();
+            log.debug("log message{}", "person was received: " + orig);
+            shower.loverUpdate(lover, orig);
+        }
         repository.save(lover);
+
         log.debug("log message{}", "person was updated: " + lover);
     }
